maze optimization

+++ Assets/Scripts/Events/PlayerHealEvent.cs
@@ -8,8 +8,8 @@ namespace Events
     public class PlayerHealEvent : EventBase
     {
         public ItemAccessbility playerType;
-        public int amount = 0;
-        public PlayerHealEvent(ItemAccessbility playerType, int amount, float delay = 0f, string name = nameof(PlayerHealEvent)) : base(name, delay)
+        public float amount = 0;
+        public PlayerHealEvent(ItemAccessbility playerType, float amount, float delay = 0f, string name = nameof(PlayerHealEvent)) : base(name, delay)
         {
             this.playerType = playerType;
             this.amount = amount;
